[opt](memory) Use Math.toIntExact for idx cast to fail-fast on overflow

+++ fe/fe-core/src/main/java/org/apache/doris/cloud/catalog/CloudReplica.java
@@ -116,7 +116,7 @@ public class CloudReplica extends Replica implements GsonPostProcessable {
         this.tableId = tableId;
         this.partitionId = partitionId;
         this.indexId = indexId;
-        this.idx = (int) idx;
+        this.idx = Math.toIntExact(idx);
     }
 
     private boolean isColocated() {
